Remove Folded Over from Device Posture
Remove the Folded Over posture because it has 2 meanings as described.
The document references "tent-mode" which typically means the device is propped up like a tent.   Tent mode does not require a secondary display, consider a Samsung Galaxy Fold that is propped open a little like an alarm clock.

A folded over position would be very challenging for developers to support generally.  If it is implemented consistently with the other APIs this means that developers will need to manually split and transform their content.