editorial: Use navigables, not browsing contexts, in CSS media feature
Related to #104.

The end result is the same, but we now avoid using deprecated terms. The
text has also been simplified: saying that something applies to the
top-level browsing context and all child contexts is redundant (it basically
says it applies to all pages); the relevant point which is that navigables
that are not top-level traversables reflect the latter's value, has been
kept.