editorial: Reorganize internal slots into different sections
In #152, it was reported that the `[[PostureOverride]]` slot was confusing,
and readers could end up thinking that its `null` value could be exposed to
script, which is not the case (this internal slot is only used internally in
calculations that result in a DevicePostureType).

To avoid confusion:
- Rename the "Internal slots" section to "Extensions to the Document
  interface", which is similar to what we already have for the Navigator
  interface.
- Move the `[[PostureOverride]]` slot's definition and the addition of
  internal slots to top-level traversables to the Automation section, and
  add a sentence explaining the values are not exposed to script directly.

While here, do some proof-reading in the "Extensions to the Document
interface" section.

Fixes #152